requirements.txt
openai>=1.13.3
pyaudio==0.2.14
pillow==10.1.0
requests==2.31.0
pyannote.audio>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
openai-whisper>=20240930
numba>=0.58.1
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=14.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
ruff>=0.0.270
mypy>=1.3.0 